`compaction`: move `enhance_key()` back to `storage` layer
To ensure that we don't use dead code in any new `compaction`
related implementations.
